Revert "Extend smoke tests with check observations"
This reverts commit 2474c5b2efc30da5a271cabadc3a8d67fe73bc86.
