+++ hydra-cluster/hydra-cluster.cabal
@@ -75,7 +75,6 @@ library
     Hydra.Cluster.Faucet
     Hydra.Cluster.Fixture
     Hydra.Cluster.Mithril
-    Hydra.Cluster.Observations
     Hydra.Cluster.Options
     Hydra.Cluster.Scenarios
     Hydra.Cluster.Util
